# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kebelet Open API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/point": {
            "get": {
                "tags": [
                    "Point"
                ],
                "summary": "Leaderboard",
                "description": "Leaderboard here",
                "operationId": "Leaderboard",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "description": "size",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "filter sort by",
                        "required": false
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "filter sort order",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfuly retrieved leaderboard",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Point"
                ],
                "summary": "Claim Point",
                "description": "Claim Point here",
                "operationId": "Claim Point",
                "responses": {
                    "200": {
                        "description": "Successfuly claim token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/point/{id}": {
            "delete": {
                "tags": [
                    "Point"
                ],
                "summary": "Delete Point",
                "description": "Delete Point here",
                "operationId": "Delete Point",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user_id",
                        "required": true
                    },
                    {
                        "name": "points",
                        "in": "query",
                        "description": "Number of points to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted points",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/point-histories": {
            "get": {
                "tags": [
                    "Point"
                ],
                "summary": "Histories point",
                "description": "Histories point here",
                "operationId": "Histories point",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "description": "size",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "filter user_id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfuly retrieved leaderboard",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Profile",
                "description": "Profile User",
                "operationId": "show-profile",
                "responses": {
                    "200": {
                        "description": "Toilet retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/{id}": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Edit Profile",
                "description": "Edit Profile here",
                "operationId": "edit profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "old_password": {
                                        "type": "string"
                                    },
                                    "profile_picture": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review successfully created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}